requirements.txt
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.15.0
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.1.5
django-allauth==0.46.0
django-cors-headers==3.6.0
django-crontab==0.7.1
django-filter==21.1
django-ipware==4.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.1
packaging==20.8
psycopg2==2.8.4
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==1.5.2
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.1
uWSGI==2.0.19.1
Werkzeug==1.0.1
whitenoise==5.2.0
